Change type of the led pin as it should be a byte instead of an int.

+++ src/01-blinky/01-blinky.ino
@@ -1,5 +1,5 @@
 // Arduino pin number of the blue LED
-const int ledPin = A12;
+const byte ledPin = A12;
 
 void setup() {
   // Set the LED pin to OUTPUT as we want to provide a signal to the
